chainlit_app/requirements.txt
chainlit
websockets
python-dotenv